+++ agent/bindings/kotlin/lib/src/main/kotlin/com/sourcegraph/cody/agent/protocol_generated/ClientCapabilities.kt
@@ -18,6 +18,7 @@ data class ClientCapabilities(
   val codeActions: CodeActionsEnum? = null, // Oneof: none, enabled
   val webviewMessages: WebviewMessagesEnum? = null, // Oneof: object-encoded, string-encoded
   val globalState: GlobalStateEnum? = null, // Oneof: stateless, server-managed, client-managed
+  val secrets: SecretsEnum? = null, // Oneof: stateless, client-managed
   val webview: WebviewEnum? = null, // Oneof: agentic, native
   val uriSchemeLoaders: List<String>? = null,
   val webviewNativeConfig: WebviewNativeConfig? = null,
@@ -90,6 +91,11 @@ data class ClientCapabilities(
   enum class GlobalStateEnum {
     @SerializedName("stateless") Stateless,
     @SerializedName("server-managed") `Server-managed`,
+    @SerializedName("client-managed") `Client-managed`,
+  }
+
+  enum class SecretsEnum {
+    @SerializedName("stateless") Stateless,
     @SerializedName("client-managed") `Client-managed`,
   }