+++ lib/shared/src/configuration.ts
@@ -1,3 +1,4 @@
+import type { Observable } from 'observable-fns'
 import type { EmbeddingsProvider } from './codebase-context/context-status'
 import type { FileURI } from './common/uri'
 
@@ -16,8 +17,13 @@ export const CONTEXT_SELECTION_ID: Record<ConfigurationUseContext, number> = {
     unified: 11,
 }
 
-export interface ConfigGetter<T> {
-    get<T>(section: string, defaultValue?: T): T
+/**
+ * A wrapper around a configuration source that lets the client retrieve the current config and
+ * watch for changes.
+ */
+export interface ConfigWatcher<C> {
+    changes: Observable<C>
+    get(): C
 }
 
 /**
@@ -31,7 +37,7 @@ export interface ClientConfiguration {
     telemetryLevel: 'all' | 'off' | 'agent'
     telemetryClientName?: string
     useContext: ConfigurationUseContext
-    customHeaders: Record<string, string>
+    customHeaders?: Record<string, string>
     chatPreInstruction: PromptString
     editPreInstruction: PromptString
     codeActions: boolean