+++ vscode/src/completions/providers/create-provider.ts
@@ -1,7 +1,7 @@
 import {
     type AuthStatus,
+    type ClientConfigurationWithAccessToken,
     type CodeCompletionsClient,
-    type ConfigurationWithAccessToken,
     FeatureFlag,
     type Model,
     ModelUsage,
@@ -21,10 +21,10 @@ import { createProviderConfig as createExperimentalOllamaProviderConfig } from '
 import { createProviderConfig as createExperimentalOpenAICompatibleProviderConfig } from './expopenaicompatible'
 import {
     DEEPSEEK_CODER_V2_LITE_BASE,
-    DEEPSEEK_CODER_V2_LITE_BASE_WINDOW_4096,
-    DEEPSEEK_CODER_V2_LITE_BASE_WINDOW_8192,
-    DEEPSEEK_CODER_V2_LITE_BASE_WINDOW_16384,
-    DEEPSEEK_CODER_V2_LITE_BASE_WINDOW_32768,
+    DEEPSEEK_CODER_V2_LITE_BASE_DIRECT_ROUTE,
+    FIREWORKS_DEEPSEEK_7B_LANG_ALL,
+    FIREWORKS_DEEPSEEK_7B_LANG_SPECIFIC_V0,
+    FIREWORKS_DEEPSEEK_7B_LANG_SPECIFIC_V1,
     type FireworksOptions,
     createProviderConfig as createFireworksProviderConfig,
 } from './fireworks'
@@ -38,7 +38,7 @@ interface CreateConfigHelperParams {
     authStatus: AuthStatus
     modelId: string | undefined
     provider: string
-    config: ConfigurationWithAccessToken
+    config: ClientConfigurationWithAccessToken
     model?: Model
 }
 
@@ -71,7 +71,6 @@ export async function createProviderConfigHelper(
             return createFireworksProviderConfig({
                 client,
                 model: modelId ?? null,
-                timeouts: config.autocompleteTimeouts,
                 authStatus,
                 config,
                 anonymousUserID,
@@ -83,7 +82,6 @@ export async function createProviderConfigHelper(
             return createExperimentalOpenAICompatibleProviderConfig({
                 client,
                 model: modelId ?? null,
-                timeouts: config.autocompleteTimeouts,
                 authStatus,
                 config,
             })
@@ -92,7 +90,6 @@ export async function createProviderConfigHelper(
             if (model) {
                 return createOpenAICompatibleProviderConfig({
                     client,
-                    timeouts: config.autocompleteTimeouts,
                     model,
                     authStatus,
                     config,
@@ -170,17 +167,16 @@ async function resolveFIMModelExperimentFromFeatureFlags(): ReturnType<
         ),
     ])
     if (fimModelVariant1) {
-        // Variant 1: Current production model with +200msec latency to quantity the effect of latency increase while keeping same quality
-        return { provider: 'fireworks', model: DEEPSEEK_CODER_V2_LITE_BASE_WINDOW_4096 }
+        return { provider: 'fireworks', model: DEEPSEEK_CODER_V2_LITE_BASE_DIRECT_ROUTE }
     }
     if (fimModelVariant2) {
-        return { provider: 'fireworks', model: DEEPSEEK_CODER_V2_LITE_BASE_WINDOW_8192 }
+        return { provider: 'fireworks', model: FIREWORKS_DEEPSEEK_7B_LANG_SPECIFIC_V0 }
     }
     if (fimModelVariant3) {
-        return { provider: 'fireworks', model: DEEPSEEK_CODER_V2_LITE_BASE_WINDOW_16384 }
+        return { provider: 'fireworks', model: FIREWORKS_DEEPSEEK_7B_LANG_SPECIFIC_V1 }
     }
     if (fimModelVariant4) {
-        return { provider: 'fireworks', model: DEEPSEEK_CODER_V2_LITE_BASE_WINDOW_32768 }
+        return { provider: 'fireworks', model: FIREWORKS_DEEPSEEK_7B_LANG_ALL }
     }
     if (fimModelCurrentBest) {
         return { provider: 'fireworks', model: DEEPSEEK_CODER_V2_LITE_BASE }
@@ -311,7 +307,7 @@ function parseProviderAndModel({
 }
 
 export async function createProviderConfig(
-    config: ConfigurationWithAccessToken,
+    config: ClientConfigurationWithAccessToken,
     client: CodeCompletionsClient,
     authStatus: AuthStatus
 ): Promise<ProviderConfig | null> {