+++ vscode/src/configuration.test.ts
@@ -1,7 +1,7 @@
 import { describe, expect, it } from 'vitest'
 import type * as vscode from 'vscode'
 
-import { type Configuration, OLLAMA_DEFAULT_URL, ps } from '@sourcegraph/cody-shared'
+import { type ClientConfiguration, OLLAMA_DEFAULT_URL, ps } from '@sourcegraph/cody-shared'
 
 import { getConfiguration } from './configuration'
 import { DEFAULT_VSCODE_SETTINGS } from './testutils/mocks'
@@ -102,8 +102,6 @@ describe('getConfiguration', () => {
                         return 1500
                     case 'cody.autocomplete.experimental.preloadDebounceInterval':
                         return 0
-                    case 'cody.autocomplete.experimental.hotStreakAndSmartThrottle':
-                        return false
                     case 'cody.experimental.guardrailsTimeoutSeconds':
                         return undefined
                     default:
@@ -151,15 +149,10 @@ describe('getConfiguration', () => {
                 model: 'codellama:7b-code',
                 url: OLLAMA_DEFAULT_URL,
             },
-            autocompleteTimeouts: {
-                multiline: undefined,
-                singleline: undefined,
-            },
             autocompleteFirstCompletionTimeout: 1500,
             autocompleteExperimentalPreloadDebounceInterval: 0,
-            autocompleteExperimentalHotStreakAndSmartThrottle: false,
             testingModelConfig: undefined,
             experimentalGuardrailsTimeoutSeconds: undefined,
-        } satisfies Configuration)
+        } satisfies ClientConfiguration)
     })
 })