+++ vscode/src/testutils/mocks.ts
@@ -10,7 +10,7 @@ import type {
 } from 'vscode'
 
 import {
-    type Configuration,
+    type ClientConfiguration,
     type FeatureFlag,
     FeatureFlagProvider,
     OLLAMA_DEFAULT_URL,
@@ -926,13 +926,8 @@ export const DEFAULT_VSCODE_SETTINGS = {
         model: 'codellama:7b-code',
         url: OLLAMA_DEFAULT_URL,
     },
-    autocompleteTimeouts: {
-        multiline: undefined,
-        singleline: undefined,
-    },
     autocompleteFirstCompletionTimeout: 3500,
     autocompleteExperimentalPreloadDebounceInterval: 0,
-    autocompleteExperimentalHotStreakAndSmartThrottle: false,
     testingModelConfig: undefined,
     experimentalGuardrailsTimeoutSeconds: undefined,
-} satisfies Configuration
+} satisfies ClientConfiguration